backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
pydantic
pydantic-settings
python-multipart
python-dotenv
requests
datasets
